refactor: move JS exports to `css-class` feature
